Created separate component for scroll button
