+++ client/src/actions/index.js
@@ -1,1 +1,3 @@
 export * from "./fetch";
+export * from "./cart-actions";
+export * from "./ui-actions";
