+++ client/src/actions/ui-actions.js
@@ -1,0 +1,4 @@
+export const toggleScrollUpButton = (bool) => ({
+  type: "TOGGLE_SCROLL_BUTTON",
+  bool,
+});
