+++ client/src/components/UI/PageContainer/PageContainer.js
@@ -1,24 +1,22 @@
 import React from "react";
 import styled from "styled-components";
-import { FiChevronUp } from "react-icons/fi";
-import { THEMES } from "../../THEMES";
+import ScrollUpButton from "../ScrollUpButton";
+import { useSelector, useDispatch } from "react-redux";
+import { toggleScrollUpButton } from "../../../actions";
 
 const PageContainer = ({ children }) => {
-  const [isVisible, setIsVisible] = React.useState(false);
+  const dispatch = useDispatch();
+
+  const ToggleScrollButton = useSelector(
+    (state) => state.TOGGLERS.ScrollButton
+  );
 
   function toggleVisibility() {
-    if (window.scrollY > 500) {
-      setIsVisible(true);
+    if (window.scrollY > 700) {
+      dispatch(toggleScrollUpButton(true));
     } else {
-      setIsVisible(false);
+      dispatch(toggleScrollUpButton(false));
     }
-  }
-
-  function scrollToTop() {
-    window.scrollTo({
-      top: 0,
-      behavior: "smooth",
-    });
   }
 
   React.useEffect(() => {
@@ -27,12 +25,7 @@ const PageContainer = ({ children }) => {
 
   return (
     <Wrapper>
-      {isVisible && (
-        <ScrollUpBtn onClick={scrollToTop}>
-          <StyledArrow size={32} />
-          <ScrollUpText>Back to top</ScrollUpText>
-        </ScrollUpBtn>
-      )}
+      {ToggleScrollButton && <ScrollUpButton />}
       <Container>{children}</Container>
     </Wrapper>
   );
@@ -40,6 +33,7 @@ const PageContainer = ({ children }) => {
 
 const Wrapper = styled.div`
   position: relative;
+  margin-top: 120px;
 `;
 
 const Container = styled.div`
@@ -67,68 +61,6 @@ const Container = styled.div`
     text-align: center;
     margin-right: 10px;
     margin-left: 10px;
-  }
-`;
-
-const StyledArrow = styled(FiChevronUp)`
-  color: white;
-`;
-
-const ScrollUpText = styled.span`
-  display: none;
-  font-size: 16px;
-  color: white;
-  font-weight: 600;
-`;
-
-const ScrollUpBtn = styled.button`
-  position: sticky;
-  z-index: 500;
-  top: 90vh;
-  left: 3%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  border-radius: 50px 50px 50px 50px;
-  border: 1px solid ${THEMES.Primary};
-  outline: none;
-  background-color: ${THEMES.Primary};
-  width: 50px;
-  height: 50px;
-  transition: 0.2s;
-  /* transform: all ease-in-out 0.3s; */
-  cursor: pointer;
-
-  &:hover {
-    position: sticky;
-    border-radius: 50px 50px 50px 50px;
-    width: 10em;
-    & ${ScrollUpText} {
-      display: inline-block;
-    }
-
-    @media (max-width: 1200px) {
-      &:hover {
-        position: sticky;
-        border-radius: 50px 50px 50px 50px;
-        width: auto;
-      }
-      & ${ScrollUpText} {
-        display: none;
-      }
-    }
-  }
-
-  @media (max-width: 1200px) {
-    /*  */
-  }
-
-  @media (max-width: 1024px) {
-    /*  */
-  }
-  @media (max-width: 768px) {
-    top: 90vh;
-    left: 6%;
   }
 `;
