+++ client/src/components/UI/ScrollUpButton/index.js
@@ -1,0 +1,1 @@
+export { default } from "./ScrollUpButton";
